+++ CHANGELOG.md
@@ -189,6 +189,10 @@
   code action even when not explicitly hovering an inexhaustive case expression.
   ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
 
+- Fixed a bug where the language server would suggest the "unqualify" code
+  action even when not explicitly hovering a qualified value.
+  ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
+
 - Fixed a bug where enabling `javascript.typescript_declarations` or
   `javascript.source_maps` wouldn't generate their additional files unless the
   build directory was manually deleted. The compiler now automatically rebuilds